remove astropy.units from general.py
